backend/requirements.txt
<<<<<<< HEAD
alchemyjsonschema==0.7.1
anyio==3.6.2
apsw==3.40.1.0
attrs==22.2.0
bcrypt==4.0.1
boto3==1.28.40
botocore==1.31.40
certifi==2022.12.7
charset-normalizer==3.0.1
click==8.1.3
configparser==5.3.0
contourpy==1.1.0
cycler==0.11.0
dictknife==0.14.0
dnspython==2.3.0
ecdsa==0.18.0
email-validator==2.2.0
et-xmlfile==1.1.0
exceptiongroup==1.1.0
fastapi==0.116.1
fastapi-pagination==0.14.0
fonttools==4.42.1
graphviz==0.20.1
greenlet==2.0.2
gunicorn==20.1.0
h11==0.14.0
idna==3.4
iniconfig==2.0.0
isodate==0.6.1
jmespath==1.0.1
jsonschema==4.17.3
kiwisolver==1.4.4
magicalimport==0.9.1
matplotlib==3.10.5
mysql-connector==2.2.9
mysql-connector-python==9.0.0
mysqlclient==2.1.1
numpy==2.2.6
opencv-python-headless==4.7.0.72
openpyxl==3.1.2
packaging==23.0
pandas==2.3.2
passlib==1.7.4
Pillow==10.0.0
pluggy==1.0.0
psycopg2-binary==2.9.6
pyasn1==0.4.8
pydantic==2.11.7
pydot==1.4.2
pymssql==2.2.8
PyMySQL==1.0.2
pyparsing==3.0.9
pyrsistent==0.19.3
pytest==7.2.2
python-dateutil==2.8.2
python-dotenv==1.0.0
python-jose==3.3.0
python-multipart==0.0.6
pytz==2022.7.1
requests==2.28.2
rsa==4.9
s3transfer==0.6.2
shillelagh==1.2.0
six==1.16.0
sniffio==1.3.0
SQLAlchemy==1.4.54
SQLAlchemy-Utils==0.41.1
sqlalchemy-schemadisplay==1.3
strict-rfc3339==0.7
tomli==2.0.1
typing_extensions>=4.12.2
urllib3==1.26.14
uvicorn==0.20.0
tabula==1.0.5
reportlab==4.0.7
weasyprint==60.1
pango==0.0.1 # sudo apt-get install -y libpangocairo-1.0-0
# bs4==0.0.1
beautifulsoup4==4.13.5
sqlalchemy_utils==0.41.1
google-analytics-data==0.18.0
# openai==1.28.0
duckdb==1.0.0
# sudo apt-get install graphviz
paramiko==3.4.1
Jinja2==3.1.3
dbt-adapters==1.2.1
dbt-common==1.3.0
dbt-core==1.8.2
dbt-duckdb==1.8.1
alembic==1.14.0
pycryptodome==3.21.0
structlog==25.1.0
psutil==5.9.6
cx_Oracle==8.3.0
oracledb==2.4.1
langchain==0.3.27
langchain-community==0.3.27
langchain-openai==0.3.31
=======
fastapi==0.104.1
uvicorn[standard]==0.24.0
alembic==1.12.1
sqlalchemy==2.0.23
pydantic==2.5.0
pydantic-settings==2.1.0
python-multipart==0.0.6
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-dotenv==1.0.0
psycopg2-binary==2.9.9
redis==5.0.1
pytest==7.4.3
pytest-asyncio==0.21.1
httpx==0.25.2
ruff==0.1.6
>>>>>>> f51b7047a2043655136d40320d507721426366cb
